Add custom tool to update VI descriptions
This tool was copied from https://gibhub.com/logmanoriginal/lvssh2 and
modified for the current project. It loads VI descriptions from the
'Descriptions.xml' file in the root directory and applies the titles and
descriptions to VIs.

Descriptions.xml contains the currend descriptions of all assertions and
the global configuration.