Change all VIMs into standard VIs
While the current approach with VIMs avoids coercion dots on the block
diagram of the caller, it does not improve the functional behavior and
introduces lookup issues in earlier versions of LabVIEW (e.g., 2017)
where VIMs are not fully discoverable in Quick Drop and other helpers.

In the future, we might want to use VIMs to utilize LabVIEW asserts to
enforce data type constraints at compile-time. This, however, requires
LabVIEW 2019 at the minimum.

This makes the public API independent of VIMs, which also enables PPL
support (currently not considered).