Major refactoring to reduce library nesting
This is a major refactoring of Transparent Assertions as a whole, to
reduce the nesting of libraries. This also fixes an issue where the
library would be loaded in- and out of memory when a caller is opened
and closed. This happens because the top-level library did not contain
any VIs and LabVIEW does not seem to be able to keep it in memory unless
a caller has a direct dependency.

The new design uses only one top-level library with all sublibraries
merged into the top-level directory (assertions, constraints,
frameworks, etc.). These files are only organized in virtual folders for
maintainability reasons.

Since files are now located in a less nested directory structure, this
also avoids issues with long path names.

This also includes changes to the test libraries to mirror the changes
introduced to the core library.