Clean up package configuration and add pipeline configuration
This removes all non-essential package dependencies from the build
specification and adds a dedicated package configuration for the build
pipeline.

This is necessary because Transparent Assertions supports different
assertion frameworks, which do not have to be installed by all users. In
the build pipeline, however, all supported frameworks must be available.