Fix Assert Collection Contains does not work for typedefs
When Assert Collection Contains is used on an array of typedef elements,
the assertion always fails even when the asserted item is in the 
collection. This happens because, when the collection is extracted into 
a variant array, typedef information is lost and LabVIEW determines that
the values are different.

The fix is to also remove typedef information from the asserted item so 
that both the array and item are of pure types. Unfortunately, it seems 
that there is no way to maintain typedef information unless we attempt 
to cast the array to an array of the asserted item, which may fail when 
the caller provides incompatible types. Breaking typedef on item is less
problematic because it does not require matching types.